+++ ChangeLog.md
@@ -14,6 +14,7 @@
     - copyDir(), copyFile() and removeTree() functions are moved to TTP::Path module
     - ttp.pl copydirs: deprecate (though still honors) --dirs option
     - ttp.pl push makes internally use of ttp.pl copydirs
+    - ttp.pl pull makes internally use of ttp.pl copydirs
 
 ### 4.1.3